1、新增：全新个人详情页<br/> 2、新增：提供修改帖子功能的web页面入口（帖子右上角）<br/> 3、优化：优化表情面板的显示效果<br/> 4、修复：修复获取任务失败的bug<br/>

+++ app/src/main/java/com/scatl/uestcbbs/entity/TaskBean.java
@@ -12,6 +12,6 @@ public class TaskBean {
     public int popularNum;
     public String award;
     public String name;
-    public int progress;
+    public double progress;
     public String icon;
 }
